+++ lib/celluloid/actor/manager.rb
@@ -1,0 +1,8 @@
+module Celluloid
+  class Actor
+    class Manager
+      include Celluloid
+      
+    end
+  end
+end
