+++ lib/celluloid/managed.rb
@@ -1,0 +1,3 @@
+$CELLULOID_BACKPORTED = false
+$CELLULOID_MANAGED = true
+require "celluloid/autostart"
